requirements.txt
asgiref==3.9.1
Django==5.2.6
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
pillow-heif==0.18.0
psycopg==3.2.10
psycopg-binary==3.2.10
sqlparse==0.5.3
